+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.1.1-dev0
+
+### Enhancements
+
+* **Added migration for GitHub Source V2**
+  * **Updated the expected output for GitHub Source V2**
+
+
 ## 0.1.0
 
 ### Enhancements